Update model definitions
