+++ db_client/alembic/versions/0067_add_indexes_for_geo_query_columns.py
@@ -1,0 +1,43 @@
+"""Add indexes for geo query columns
+
+Revision ID: 0067
+Revises: 0066
+Create Date: 2025-04-30 09:50:22.300653
+
+"""
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision = "0067"
+down_revision = "0066"
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_index(op.f("ix_corpus_import_id"), "corpus", ["import_id"], unique=True)
+    op.create_index(op.f("ix_family_import_id"), "family", ["import_id"], unique=True)
+    op.create_index(
+        op.f("ix_family_corpus_corpus_import_id"),
+        "family_corpus",
+        ["corpus_import_id"],
+        unique=False,
+    )
+    op.create_index(
+        op.f("ix_family_corpus_family_import_id"),
+        "family_corpus",
+        ["family_import_id"],
+        unique=False,
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f("ix_family_corpus_family_import_id"), table_name="family_corpus")
+    op.drop_index(op.f("ix_family_corpus_corpus_import_id"), table_name="family_corpus")
+    op.drop_index(op.f("ix_family_import_id"), table_name="family")
+    op.drop_index(op.f("ix_corpus_import_id"), table_name="corpus")
+    # ### end Alembic commands ###
